fix(B-464): split conversation scenario model and schema
