+++ backend/app/models/__init__.py
@@ -3,8 +3,6 @@ from app.models.app_config import AppConfig, ConfigType
 from app.models.conversation_category import ConversationCategory
 from app.models.conversation_scenario import (
     ConversationScenario,
-    ConversationScenarioCreate,
-    ConversationScenarioRead,
     ConversationScenarioStatus,
 )
 from app.models.personalization_option import PersonalizationOptionRead
@@ -52,8 +50,6 @@ from app.models.user_profile import (
 __all__ = [
     'ConversationCategory',
     'ConversationScenario',
-    'ConversationScenarioCreate',
-    'ConversationScenarioRead',
     'ConversationScenarioStatus',
     'Session',
     'SessionCreate',
